+++ src/Emails.php
@@ -9,6 +9,7 @@ use Ryssbowh\CraftEmails\emailSources\AllUsersEmailSource;
 use Ryssbowh\CraftEmails\emailSources\MailchimpEmailSource;
 use Ryssbowh\CraftEmails\emailSources\UserGroupEmailSource;
 use Ryssbowh\CraftEmails\events\RegisterEmailSourcesEvent;
+use Ryssbowh\CraftEmails\jobs\ReinstallJob;
 use Ryssbowh\CraftEmails\models\Settings;
 use Ryssbowh\CraftEmails\models\actions\SendEmail;
 use Ryssbowh\CraftEmails\services\AttachementsService;
@@ -33,7 +34,7 @@ use craft\events\RegisterTemplateRootsEvent;
 use craft\events\RegisterUrlRulesEvent;
 use craft\events\RegisterUserPermissionsEvent;
 use craft\helpers\App;
-use craft\helpers\FileHelper;
+use craft\helpers\Queue;
 use craft\mail\Mailer;
 use craft\models\SystemMessage;
 use craft\records\Site;
@@ -101,6 +102,7 @@ class Emails extends Plugin
         $this->registerSiteChange();
         $this->registerBehaviors();
         $this->registerTriggers();
+        $this->registerPluginEvents();
 
         if (Craft::$app->request->getIsCpRequest()) {
             $this->registerCpRoutes();
@@ -140,6 +142,23 @@ class Emails extends Plugin
             return $item;
         }
         return null;
+    }
+
+    /**
+     * Register plugins events
+     *
+     * @since 2.0.8
+     */
+    protected function registerPluginEvents()
+    {
+        Event::on(Plugins::class, Plugins::EVENT_AFTER_INSTALL_PLUGIN, function () {
+            Emails::$plugin->emails->install();
+        });
+        Event::on(Plugins::class, Plugins::EVENT_AFTER_UNINSTALL_PLUGIN, function (Event $e) {
+            if ($e->plugin->handle !== 'emails') {
+                Queue::push(new ReinstallJob);
+            }
+        });
     }
 
     /**
@@ -252,7 +271,7 @@ class Emails extends Plugin
                 foreach (Emails::$plugin->mailchimp->lists as $list) {
                     $e->add(new MailchimpEmailSource([
                         'id' => $list['id']
-                    ]));   
+                    ]));
                 }
             }
         );
@@ -390,27 +409,25 @@ class Emails extends Plugin
     }
 
     /**
-     * After theme is installed, creates system emails.
-     */
-    protected function afterInstall(): void
-    {
-        Event::on(
-            Plugins::class,
-            Plugins::EVENT_AFTER_INSTALL_PLUGIN,
-            function () {
-                Emails::$plugin->emails->install();
-            }
-        );
-    }
-
-    /**
-     * Remove all config after uninstall
+     * @inheritdoc
      */
     protected function afterUninstall(): void
     {
-        Craft::$app->getProjectConfig()->remove('emails');
+        if (\Craft::$app->projectConfig->get(EmailsService::CONFIG_KEY, true)) {
+            Craft::$app->getProjectConfig()->remove(EmailsService::CONFIG_KEY);
+        }
         \Craft::$app->getDb()->createCommand()
             ->delete(Table::SYSTEMMESSAGES)
             ->execute();
+    }
+
+    /**
+     * @inheritdoc
+     */
+    protected function afterInstall(): void
+    {
+        if (!\Craft::$app->projectConfig->get('plugins.redactor', true)) {
+            \Craft::$app->plugins->installPlugin('redactor');
+        }
     }
 }